only compact intral sub level for small write
