+++ src/meta/src/hummock/compaction/level_selector.rs
@@ -33,9 +33,8 @@ use super::{
     create_compaction_task, LevelCompactionPicker, ManualCompactionOption, ManualCompactionPicker,
     TierCompactionPicker,
 };
-use crate::hummock::compaction::overlap_strategy::OverlapStrategy;
 use crate::hummock::compaction::picker::{
-    CompactionPicker, LocalPickerStatistic, MinOverlappingPicker,
+    CompactionPicker, IntraSubLevelPicker, LocalPickerStatistic, MinOverlappingPicker,
 };
 use crate::hummock::compaction::{create_overlap_strategy, CompactionTask, LocalSelectorStatistic};
 use crate::hummock::level_handler::LevelHandler;
@@ -62,7 +61,14 @@ pub trait LevelSelector: Sync + Send {
     fn task_type(&self) -> compact_task::TaskType;
 }
 
-#[derive(Default, Debug)]
+pub struct PickerInfo {
+    score: u64,
+    select_level: usize,
+    target_level: usize,
+    picker: Box<dyn CompactionPicker>,
+}
+
+#[derive(Default)]
 pub struct SelectContext {
     pub level_max_bytes: Vec<u64>,
 
@@ -71,7 +77,7 @@ pub struct SelectContext {
     // size of the files in  `base_level` reaches its capacity, we will place data in a higher
     // level, which equals to `base_level -= 1;`.
     pub base_level: usize,
-    pub score_levels: Vec<(u64, usize, usize)>,
+    pub score_levels: Vec<PickerInfo>,
 }
 
 pub struct DynamicLevelSelectorCore {
@@ -88,33 +94,6 @@ impl DynamicLevelSelectorCore {
 
     pub fn get_config(&self) -> &CompactionConfig {
         self.config.as_ref()
-    }
-
-    fn create_compaction_picker(
-        &self,
-        select_level: usize,
-        target_level: usize,
-        overlap_strategy: Arc<dyn OverlapStrategy>,
-    ) -> Box<dyn CompactionPicker> {
-        if select_level == 0 {
-            if target_level == 0 {
-                Box::new(TierCompactionPicker::new(self.config.clone()))
-            } else {
-                Box::new(LevelCompactionPicker::new(
-                    target_level,
-                    self.config.clone(),
-                ))
-            }
-        } else {
-            assert_eq!(select_level + 1, target_level);
-            Box::new(MinOverlappingPicker::new(
-                select_level,
-                target_level,
-                self.config.max_bytes_for_level_base,
-                self.config.split_by_state_table,
-                overlap_strategy,
-            ))
-        }
     }
 
     // TODO: calculate this scores in apply compact result.
@@ -178,7 +157,12 @@ impl DynamicLevelSelectorCore {
         ctx
     }
 
-    fn get_priority_levels(&self, levels: &Levels, handlers: &[LevelHandler]) -> SelectContext {
+    fn get_priority_levels(
+        &self,
+        levels: &Levels,
+        handlers: &[LevelHandler],
+        config: Arc<CompactionConfig>,
+    ) -> SelectContext {
         let mut ctx = self.calculate_level_base_size(levels);
 
         let idle_file_count = levels
@@ -190,7 +174,6 @@ impl DynamicLevelSelectorCore {
             .map(|level| level.table_infos.len())
             .sum::<usize>()
             - handlers[0].get_pending_file_count();
-
         if idle_file_count > 0 {
             // trigger l0 compaction when the number of files is too large.
 
@@ -212,46 +195,59 @@ impl DynamicLevelSelectorCore {
                     std::cmp::min(idle_file_count, overlapping_file_count) as u64 * SCORE_BASE
                         / self.config.level0_tier_compact_file_number;
                 // Reduce the level num of l0 overlapping sub_level
-                ctx.score_levels
-                    .push((std::cmp::max(l0_overlapping_score, SCORE_BASE + 1), 0, 0));
+                ctx.score_levels.push(PickerInfo {
+                    score: std::cmp::max(l0_overlapping_score, SCORE_BASE + 1),
+                    select_level: 0,
+                    target_level: 0,
+                    picker: Box::new(TierCompactionPicker::new(config.clone())),
+                });
             }
 
             // The read query at the non-overlapping level only selects ssts that match the query
             // range at each level, so the number of levels is the most important factor affecting
             // the read performance. At the same time, the size factor is also added to the score
             // calculation rule to avoid unbalanced compact task due to large size.
-            let non_overlapping_score = {
-                let total_size = levels.l0.as_ref().unwrap().total_file_size
-                    - handlers[0].get_pending_output_file_size(ctx.base_level as u32);
-                let base_level_size = levels.get_level(ctx.base_level).total_file_size;
-                let base_level_sst_count =
-                    levels.get_level(ctx.base_level).table_infos.len() as u64;
+            let total_size = levels.l0.as_ref().unwrap().total_file_size
+                - handlers[0].get_pending_output_file_size(ctx.base_level as u32);
+            let base_level_size = levels.get_level(ctx.base_level).total_file_size;
+            let base_level_sst_count = levels.get_level(ctx.base_level).table_infos.len() as u64;
 
-                // size limit
-                let non_overlapping_size_score = total_size * SCORE_BASE
-                    / std::cmp::max(self.config.max_bytes_for_level_base, base_level_size);
-                // level count limit
-                let non_overlapping_level_count = levels
-                    .l0
-                    .as_ref()
-                    .unwrap()
-                    .sub_levels
-                    .iter()
-                    .filter(|level| level.level_type() == LevelType::Nonoverlapping)
-                    .count() as u64;
-                let non_overlapping_level_score = non_overlapping_level_count * SCORE_BASE
-                    / std::cmp::max(
-                        base_level_sst_count / 16,
-                        self.config.level0_sub_level_compact_level_count as u64,
-                    );
+            // size limit
+            let non_overlapping_size_score = total_size * SCORE_BASE
+                / std::cmp::max(self.config.max_bytes_for_level_base, base_level_size);
+            // level count limit
+            let non_overlapping_level_count = levels
+                .l0
+                .as_ref()
+                .unwrap()
+                .sub_levels
+                .iter()
+                .filter(|level| level.level_type() == LevelType::Nonoverlapping)
+                .count() as u64;
+            let non_overlapping_level_score = non_overlapping_level_count * SCORE_BASE
+                / std::cmp::max(
+                    base_level_sst_count / 16,
+                    self.config.level0_sub_level_compact_level_count as u64,
+                );
 
-                std::cmp::max(non_overlapping_size_score, non_overlapping_level_score)
-            };
-
-            // Reduce the level num of l0 non-overlapping sub_level
-            ctx.score_levels
-                .push((non_overlapping_score, 0, ctx.base_level));
+            if non_overlapping_size_score > SCORE_BASE {
+                // Reduce the level num of l0 non-overlapping sub_level
+                ctx.score_levels.push(PickerInfo {
+                    score: non_overlapping_size_score,
+                    select_level: 0,
+                    target_level: ctx.base_level,
+                    picker: Box::new(LevelCompactionPicker::new(ctx.base_level, config.clone())),
+                });
+            } else if non_overlapping_level_score > SCORE_BASE {
+                ctx.score_levels.push(PickerInfo {
+                    score: non_overlapping_level_score,
+                    select_level: 0,
+                    target_level: ctx.base_level,
+                    picker: Box::new(IntraSubLevelPicker::new(config.clone())),
+                });
+            }
         }
+        let overlap_strategy = create_overlap_strategy(config.compaction_mode());
 
         // The bottommost level can not be input level.
         for level in &levels.levels {
@@ -270,16 +266,28 @@ impl DynamicLevelSelectorCore {
             if total_size == 0 {
                 continue;
             }
-            ctx.score_levels.push((
-                total_size * SCORE_BASE / ctx.level_max_bytes[level_idx],
-                level_idx,
-                level_idx + 1,
-            ));
+            let score = total_size * SCORE_BASE / ctx.level_max_bytes[level_idx];
+            if score >= SCORE_BASE {
+                ctx.score_levels.push(PickerInfo {
+                    score,
+                    select_level: level_idx,
+                    target_level: level_idx + 1,
+                    picker: Box::new(MinOverlappingPicker::new(
+                        level_idx,
+                        level_idx + 1,
+                        self.config.max_bytes_for_level_base,
+                        self.config.split_by_state_table,
+                        overlap_strategy.clone(),
+                    )),
+                });
+            }
         }
-
         // sort reverse to pick the largest one.
-        ctx.score_levels
-            .sort_by(|a, b| b.0.cmp(&a.0).then_with(|| a.2.cmp(&b.2)));
+        ctx.score_levels.sort_by(|a, b| {
+            b.score
+                .cmp(&a.score)
+                .then_with(|| a.target_level.cmp(&b.target_level))
+        });
         ctx
     }
 
@@ -371,20 +379,19 @@ impl LevelSelector for DynamicLevelSelector {
     ) -> Option<CompactionTask> {
         let dynamic_level_core =
             DynamicLevelSelectorCore::new(compaction_group.compaction_config.clone());
-        let overlap_strategy =
-            create_overlap_strategy(compaction_group.compaction_config.compaction_mode());
-        let ctx = dynamic_level_core.get_priority_levels(levels, level_handlers);
-        for (score, select_level, target_level) in ctx.score_levels {
-            if score <= SCORE_BASE {
-                return None;
-            }
-            let mut picker = dynamic_level_core.create_compaction_picker(
-                select_level,
-                target_level,
-                overlap_strategy.clone(),
-            );
+
+        let ctx = dynamic_level_core.get_priority_levels(
+            levels,
+            level_handlers,
+            compaction_group.compaction_config.clone(),
+        );
+        for mut picker_info in ctx.score_levels {
             let mut stats = LocalPickerStatistic::default();
-            if let Some(ret) = picker.pick_compaction(levels, level_handlers, &mut stats) {
+            if let Some(ret) =
+                picker_info
+                    .picker
+                    .pick_compaction(levels, level_handlers, &mut stats)
+            {
                 ret.add_pending_task(task_id, level_handlers);
                 return Some(create_compaction_task(
                     dynamic_level_core.get_config(),
@@ -393,9 +400,11 @@ impl LevelSelector for DynamicLevelSelector {
                     self.task_type(),
                 ));
             }
-            selector_stats
-                .skip_picker
-                .push((select_level, target_level, stats));
+            selector_stats.skip_picker.push((
+                picker_info.select_level,
+                picker_info.target_level,
+                stats,
+            ));
         }
         None
     }
